Create observability-book
Hey, hi there. This is for a forked repo from Accenture's Udacity. An e-learning space. A good one too, but it's too expensive for me to afford right now. Hopefully... if God is willing, in the future, it will also become possible for me. 

Anyway, I'm going to set up a K8s sandbox. And will start implementing observability helpers. Thanks.